Rename util to fmt.
