+++ src/broadcast.rs
@@ -1,3 +1,4 @@
+use fmt::{HexBytes, HexList, HexProof};
 use merkle::proof::{Lemma, Positioned, Proof};
 use merkle::MerkleTree;
 use reed_solomon_erasure as rse;
@@ -9,7 +10,6 @@ use std::collections::{BTreeMap, BTreeSet, VecDeque};
 use std::fmt::{self, Debug};
 use std::hash::Hash;
 use std::iter;
-use util::{HexBytes, HexList, HexProof};
 
 use messaging::{Target, TargetedMessage};
